+++ src/main/java/com/example/ecommerceweb/controller/CheckOutController.java
@@ -1,9 +1,8 @@
 package com.example.ecommerceweb.controller;
 
-import com.example.ecommerceweb.dto.request.CheckoutRequest;
-import com.example.ecommerceweb.dto.request.ProductRequest;
-import com.example.ecommerceweb.dto.response.StripeResponse;
-import com.example.ecommerceweb.service.PaymentService;
+import com.example.ecommerceweb.dto.request.order.CheckoutRequest;
+import com.example.ecommerceweb.dto.response.order.StripeResponse;
+import com.example.ecommerceweb.service.services_impl.PaymentService;
 import com.stripe.exception.StripeException;
 import com.stripe.model.PaymentIntent;
 import lombok.RequiredArgsConstructor;
@@ -19,22 +18,11 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class CheckOutController {
 
-    // TODO: build api for checkout
-
     private final PaymentService paymentService;
-
 
     @PostMapping("/create-payment-intent")
     public PaymentIntent createPayment(@RequestParam Long amount, @RequestParam String currency) throws StripeException {
         return paymentService.createPaymentIntent(amount, currency);
-    }
-
-    @PostMapping("/checkout-order")
-    public ResponseEntity<StripeResponse> checkoutProducts(@RequestBody ProductRequest productRequest) {
-        StripeResponse stripeResponse = paymentService.checkoutProducts(productRequest);
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(stripeResponse);
     }
 
     @PostMapping("/checkout-list")
